Improve Universal Dependencies support
Addresses (most of) issue #10
(does not handle the feature mappings currently)

StanfordDependencies: new "universal" parameter to convert_tree()
which outputs UD if available (otherwise outputs SD with potentially no
warnings since SubprocessBackend can't know from an arbitrary jar
whether it is new enough to support UD). To match CoreNLP behavior,
universal defaults to True. Added and updated documentation.

JPypeBackend: checks for UD support and enables it if requested and available (unlike SubprocessBackend, this gives a warning if not available)

SubprocessBackend: minimal updates to support UD

Testing code: String data was (finally) pulled out of
test_stanforddependencies and moved to test/data.py.
The strings needed to be generalized to include UD variants
of the output, so a class structure is used. We now test
SD and UD versions of both backends by extending the
DefaultBackendTest hierarchy.